backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
deepgram-sdk
openai
chromadb
tiktoken
